pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "47chat"
dynamic = ["version"]
description = "AI Multi-Agent Orchestrator with RAG capabilities"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "47Chat Contributors", email = "contributors@47chat.dev"}
]
maintainers = [
    {name = "47Chat Contributors", email = "maintainers@47chat.dev"}
]
keywords = [
    "ai",
    "multi-agent",
    "orchestrator",
    "rag",
    "llm",
    "fastapi",
    "streamlit",
    "ollama",
    "openai",
    "machine-learning",
    "nlp",
    "chatbot",
    "automation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Framework :: FastAPI",
    "Framework :: AsyncIO",
]
urls = {Repository = "https://github.com/Hipson47/47Chat.git"}
dependencies = [
    "fastapi>=0.104,<1.0",
    "uvicorn[standard]>=0.24,<1.0",
    "python-multipart>=0.0.6,<1.0",
    "faiss-cpu>=1.7,<2.0",
    "sentence-transformers>=2.2,<3.0",
    "pypdf>=3.17,<4.0",
    "markdown>=3.5,<4.0",
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=4.9,<5.0",
    "requests>=2.31,<3.0",
    "pydantic>=2.5,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "pyyaml>=6.0,<7.0",
    "streamlit>=1.29,<2.0",
    "langgraph>=0.0.32,<1.0",
    "slowapi>=0.1.9,<1.0",
    "openai>=1.6,<2.0"
]

[tool.setuptools]
packages = ["backend", "frontend"]

[project.optional-dependencies]
dev = [
    "ruff>=0.5,<0.6",
    "mypy>=1.11,<2",
    "pytest>=8.3,<9",
    "pytest-cov>=5,<6",
    "pytest-xdist>=3.5,<4",
    "pytest-benchmark>=4.0,<5",
    "hypothesis>=6.100,<7",
    "safety>=3,<4",
    "pre-commit>=3.5,<4",
    "structlog>=23.1,<24",
    "opentelemetry-sdk>=1.20,<2",
    "opentelemetry-instrumentation-fastapi>=0.41b0,<1",
    "opentelemetry-exporter-otlp>=1.20,<2",
    "prometheus-client>=0.19,<1",
    "python-semantic-release>=9.0,<10"
]
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0"
]
docs = [
    "sphinx>=7.0,<8.0",
    "sphinx-rtd-theme>=1.3,<2.0",
    "myst-parser>=2.0,<3.0"
]
observability = [
    "structlog>=23.1,<24",
    "opentelemetry-sdk>=1.20,<2",
    "opentelemetry-instrumentation-fastapi>=0.41b0,<1",
    "opentelemetry-exporter-otlp>=1.20,<2",
    "prometheus-client>=0.19,<1"
]
all = [
    "47chat[dev,gpu,docs,observability]"
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "W", "PL", "RUF"]
ignore = ["D"]
extend-select = ["I", "UP", "F", "E", "W"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "frontend"]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]
enable_error_code = ["redundant-cast"]
ignore_missing_imports = true
exclude = ["build/"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-q",
    "-ra",
    "-n", "auto",
    "--cov=backend",
    "--cov=frontend",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=85",
    "--disable-warnings",
    "--strict-markers",
    "--strict-config"
]
testpaths = ["tests", "backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "property: marks tests as property-based tests",
    "smoke: marks tests as smoke tests"
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = [
    "backend/__init__.py:__version__",
    "frontend/__init__.py:__version__"
]
branch = "main"
changelog_file = "CHANGELOG.md"
build_command = "python -m pip install build && python -m build"
dist_glob_patterns = ["dist/*"]
upload_to_pypi = false
upload_to_release = true
github_release = true
github_release_name_template = "Release {version} 🚀"
github_release_notes_template = """
## 🚀 Release {version}

{notes}

### 📦 Installation
```bash
pip install 47chat=={version}
```

### 🔗 Links
- [Compare Changes](https://github.com/Hipson47/47Chat/compare/{previous_version}...{version})
- [Full Changelog](https://github.com/Hipson47/47Chat/blob/main/CHANGELOG.md)
"""
commit_message = "chore(release): bump version to {version} [skip ci]"
commit_author = "semantic-release <semantic-release@47chat.dev>"

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = [
    "chore.*",
    "ci.*",
    "build.*",
    "test.*",
    "docs.*",
    "style.*",
    "refactor.*",
    "perf.*"
]

[tool.semantic_release.commit_parser_options]
major_types = ["breaking"]
minor_types = ["feat"]
patch_types = ["fix", "perf"]
other_types = ["docs", "style", "refactor", "test", "build", "ci", "chore"]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test", "breaking"]
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure_skip_verify = false
